+++ src/app/js/model.js
@@ -2,15 +2,23 @@
 Attribute-based data model with APIs for getting, setting, validating, and
 syncing attribute values, as well as events for being notified of model changes.
 
+@submodule model
+@since 3.4.0
+**/
+
+/**
+Attribute-based data model with APIs for getting, setting, validating, and
+syncing attribute values, as well as events for being notified of model changes.
+
 In most cases, you'll want to create your own subclass of `Y.Model` and
 customize it to meet your needs. In particular, the `sync()` and `validate()`
 methods are meant to be overridden by custom implementations. You may also want
 to override the `parse()` method to parse non-generic server responses.
 
-@submodule model
 @class Model
 @constructor
-@uses Base
+@extends Base
+@since 3.4.0
 **/
 
 var GlobalEnv = YUI.namespace('Env.Model'),
@@ -41,8 +49,8 @@ var GlobalEnv = YUI.namespace('Env.Model'),
     @param {String} src Source of the error. May be one of the following (or any
       custom error source defined by a Model subclass):
 
-        * `parse`: An error parsing a JSON response.
-        * `validate`: The model failed to validate.
+      * `parse`: An error parsing a JSON response.
+      * `validate`: The model failed to validate.
     **/
     EVT_ERROR = 'error';
 
@@ -161,7 +169,7 @@ Y.Model = Y.extend(Model, Y.Base, {
 
         function finish(err) {
             if (!err) {
-                YArray.each(self.lists, function (list) {
+                YArray.each(self.lists.concat(), function (list) {
                     list.remove(self, options);
                 });
 
@@ -526,7 +534,7 @@ Y.Model = Y.extend(Model, Y.Base, {
 
       * `create`: Store a newly-created model for the first time.
       * `delete`: Delete an existing model.
-      * 'read'  : Load an existing model.
+      * `read`  : Load an existing model.
       * `update`: Update an existing model.
 
     @param {Object} [options] Sync options. It's up to the custom sync