+++ src/autocomplete/js/autocomplete-base.js
@@ -11,7 +11,6 @@
  * UI implementation) for a text input field or textarea. Must be mixed into a
  * <code>Y.Base</code>-derived class to be useful.
  *
- * @module autocomplete
  * @submodule autocomplete-base
  */
 
@@ -865,7 +864,7 @@ AutoCompleteBase.prototype = {
         if (query || query === '') {
             this._set(QUERY, query);
         } else {
-            query = this.get(QUERY);
+            query = this.get(QUERY) || '';
         }
 
         if (source) {
@@ -873,7 +872,8 @@ AutoCompleteBase.prototype = {
                 requestTemplate = this.get(REQUEST_TEMPLATE);
             }
 
-            request = requestTemplate ? requestTemplate(query) : query;
+            request = requestTemplate ?
+                requestTemplate.call(this, query) : query;
 
             Y.log('sendRequest: ' + request, 'info', 'autocomplete-base');
 
@@ -1101,7 +1101,7 @@ AutoCompleteBase.prototype = {
             textLocator;
 
         if (unfiltered && listLocator) {
-            unfiltered = listLocator(unfiltered);
+            unfiltered = listLocator.call(this, unfiltered);
         }
 
         if (unfiltered && unfiltered.length) {
@@ -1114,7 +1114,10 @@ AutoCompleteBase.prototype = {
             // as we go.
             for (i = 0, len = unfiltered.length; i < len; ++i) {
                 result = unfiltered[i];
-                text   = textLocator ? textLocator(result) : result.toString();
+
+                text = textLocator ?
+                        textLocator.call(this, result) :
+                        result.toString();
 
                 results.push({
                     display: Escape.html(text),
@@ -1127,7 +1130,7 @@ AutoCompleteBase.prototype = {
             // filter returns an array of (potentially fewer) result objects,
             // which is then passed to the next filter, and so on.
             for (i = 0, len = filters.length; i < len; ++i) {
-                results = filters[i](query, results.concat());
+                results = filters[i].call(this, query, results.concat());
 
                 if (!results) {
                     Y.log("Filter didn't return anything.", 'warn', 'autocomplete-base');
@@ -1156,7 +1159,8 @@ AutoCompleteBase.prototype = {
                 // an array of result objects), and these strings are then added
                 // to each result object.
                 if (highlighter) {
-                    highlighted = highlighter(query, results.concat());
+                    highlighted = highlighter.call(this, query,
+                            results.concat());
 
                     if (!highlighted) {
                         Y.log("Highlighter didn't return anything.", 'warn', 'autocomplete-base');
@@ -1176,7 +1180,7 @@ AutoCompleteBase.prototype = {
                 // objects), and these strings/Nodes are then added to each
                 // result object.
                 if (formatter) {
-                    formatted = formatter(query, results.concat());
+                    formatted = formatter.call(this, query, results.concat());
 
                     if (!formatted) {
                         Y.log("Formatter didn't return anything.", 'warn', 'autocomplete-base');
@@ -1575,8 +1579,8 @@ AutoCompleteBase.prototype = {
      */
     _onResponse: function (query, e) {
         // Ignore stale responses that aren't for the current query.
-        if (query === this.get(QUERY)) {
-            this._parseResponse(query, e.response, e.data);
+        if (query === (this.get(QUERY) || '')) {
+            this._parseResponse(query || '', e.response, e.data);
         }
     },