+++ src/autocomplete/js/autocomplete-list.js
@@ -1,8 +1,12 @@
 /**
  * Traditional autocomplete dropdown list widget, just like Mom used to make.
  *
- * @module autocomplete
  * @submodule autocomplete-list
+ */
+
+/**
+ * Traditional autocomplete dropdown list widget, just like Mom used to make.
+ * 
  * @class AutoCompleteList
  * @extends Widget
  * @uses AutoCompleteBase
@@ -122,12 +126,8 @@ List = Y.Base.create('autocompleteList', Y.Widget, [
             boundingBox = this.get('boundingBox'),
             contentBox  = this.get('contentBox'),
             inputNode   = this._inputNode,
-            listNode,
+            listNode    = this._createListNode(),
             parentNode  = inputNode.get('parentNode');
-
-        listNode = this._createListNode();
-        this._set('listNode', listNode);
-        contentBox.append(listNode);
 
         inputNode.addClass(this.getClassName('input')).setAttrs({
             'aria-autocomplete': LIST,
@@ -352,7 +352,8 @@ List = Y.Base.create('autocompleteList', Y.Widget, [
               visibleChange       : this._afterVisibleChange
             }),
 
-            this._listNode.delegate('click', this._onItemClick, this[_SELECTOR_ITEM], this)
+            this._listNode.delegate('click', this._onItemClick,
+                    this[_SELECTOR_ITEM], this)
         ]);
     },
 
@@ -403,19 +404,25 @@ List = Y.Base.create('autocompleteList', Y.Widget, [
     },
 
     /**
-     * Creates and returns a list node.
+     * Creates and returns a list node. If the `listNode` attribute is already
+     * set to an existing node, that node will be used.
      *
      * @method _createListNode
      * @return {Node} List node.
      * @protected
      */
     _createListNode: function () {
-        var listNode = Node.create(this.LIST_TEMPLATE);
+        var listNode = this.get('listNode') || Node.create(this.LIST_TEMPLATE);
 
-        return listNode.addClass(this.getClassName(LIST)).setAttrs({
+        listNode.addClass(this.getClassName(LIST)).setAttrs({
             id  : Y.stamp(listNode),
             role: 'listbox'
         });
+
+        this._set('listNode', listNode);
+        this.get('contentBox').append(listNode);
+
+        return listNode;
     },
 
     /**
@@ -542,7 +549,8 @@ List = Y.Base.create('autocompleteList', Y.Widget, [
     _afterActiveItemChange: function (e) {
         var inputNode = this._inputNode,
             newVal    = e.newVal,
-            prevVal   = e.prevVal;
+            prevVal   = e.prevVal,
+            node;
 
         // The previous item may have disappeared by the time this handler runs,
         // so we need to be careful.
@@ -558,7 +566,13 @@ List = Y.Base.create('autocompleteList', Y.Widget, [
         }
 
         if (this.get('scrollIntoView')) {
-            (newVal || inputNode).scrollIntoView();
+            node = newVal || inputNode;
+
+            if (!node.inRegion(Y.DOM.viewportRegion(), true)
+                    || !node.inRegion(this._contentBox, true)) {
+
+                node.scrollIntoView();
+            }
         }
     },
 
@@ -752,7 +766,7 @@ List = Y.Base.create('autocompleteList', Y.Widget, [
          *
          * @attribute hoveredItem
          * @type Node|null
-         * @readonly
+         * @readOnly
          */
         hoveredItem: {
             readOnly: true,
@@ -764,10 +778,10 @@ List = Y.Base.create('autocompleteList', Y.Widget, [
          *
          * @attribute listNode
          * @type Node|null
-         * @readonly
+         * @initOnly
          */
         listNode: {
-            readOnly: true,
+            writeOnce: 'initOnly',
             value: null
         },