+++ src/autocomplete/js/autocomplete-sources.js
@@ -88,7 +88,7 @@ Y.mix(ACBase.prototype, {
                 ioRequest.abort();
             }
 
-            ioRequest = Y.io(that._getXHRUrl(source, query), {
+            ioRequest = Y.io(that._getXHRUrl(source, request), {
                 on: {
                     success: function (tid, response) {
                         var data;
@@ -392,22 +392,23 @@ Y.mix(ACBase.prototype, {
      *
      * @method _getXHRUrl
      * @param {String} url Base URL.
-     * @param {String} query AutoComplete query.
+     * @param {Object} request Request object containing `query` and `request`
+     *   properties.
      * @return {String} Formatted URL.
      * @protected
      * @for AutoCompleteBase
      */
-    _getXHRUrl: function (url, query) {
-        var maxResults      = this.get(MAX_RESULTS),
-            requestTemplate = this.get(REQUEST_TEMPLATE);
+    _getXHRUrl: function (url, request) {
+        var maxResults = this.get(MAX_RESULTS);
 
-        if (requestTemplate) {
-            url += requestTemplate(query);
+        if (request.query !== request.request) {
+            // Append the request template to the URL.
+            url += request.request;
         }
 
         return Lang.sub(url, {
             maxResults: maxResults > 0 ? maxResults : 1000,
-            query     : encodeURIComponent(query)
+            query     : encodeURIComponent(request.query)
         });
     },