+++ src/base/js/Base.js
@@ -189,7 +189,6 @@
          * invoking initializers for the class hierarchy.
          *
          * @method init
-         * @final
          * @chainable
          * @param {Object} config Object with configuration property name/value pairs
          * @return {Base} A reference to this object
@@ -277,7 +276,6 @@
          * </p>
          * @method destroy
          * @return {Base} A reference to this object
-         * @final
          * @chainable
          */
         destroy: function() {