+++ src/widget-autohide/js/Widget-Autohide.js
@@ -36,159 +36,136 @@ var WIDGET_AUTOHIDE    = 'widgetAutohide',
     RENDERED            = "rendered",
     BOUNDING_BOX        = "boundingBox",
     VISIBLE             = "visible",
-    HOST                = "host",
     CHANGE              = 'Change',
 
     getCN               = Y.ClassNameManager.getClassName;
 
+function WidgetAutohide(config) {
+    this._initAutohide();
 
-WidgetAutohide = Y.Base.create(WIDGET_AUTOHIDE, Y.Plugin.Base, [], {
+}
 
+WidgetAutohide.ATTRS = {
+    hideOn: {
+        value: [
+            // {
+            //     eventName: CLICK_OUTSIDE
+            // },
+            // {
+            //     eventName: FOCUS_OUTSIDE
+            // },
+
+            {
+                node: Y.one(DOCUMENT),
+                eventName: KEY,
+                keyCode: PRESS_ESCAPE
+            }
+        ],
+        validator: Y.Lang.isArray
+    }
+};
+
+WidgetAutohide.prototype = {
     // *** Instance Members *** //
 
-    _uiHandles : null,
+        _uiHandlesAutohide : null,
 
-    // *** Lifecycle Methods *** //
+        // *** Lifecycle Methods *** //
 
-    initializer : function (config) {
+        _initAutohide : function (config) {
+            Y.after(this._bindUIAutohide, this, BIND_UI);
+            Y.after(this._syncUIAutohide, this, SYNC_UI);
 
-        this.afterHostMethod(BIND_UI, this.bindUI);
-        this.afterHostMethod(SYNC_UI, this.syncUI);
 
-        if (this.get(HOST).get(RENDERED)) {
-            this.bindUI();
-            this.syncUI();
-        }
-    },
-
-    destructor : function () {
-
-        this._detachUIHandles();
-    },
-
-    bindUI : function () {
-
-        this.afterHostEvent(VISIBLE+CHANGE, this._afterHostVisibleChange);
-    },
-
-    syncUI : function () {
-
-        this._uiSetHostVisible(this.get(HOST).get(VISIBLE));
-    },
-
-    // *** Private Methods *** //
-
-    _uiSetHostVisible : function (visible) {
-
-        if (visible) {
-            //this._attachUIHandles();
-            Y.later(1, this, '_attachUIHandles');
-        } else {
-            this._detachUIHandles();
-        }
-    },
-
-    _attachUIHandles : function () {
-
-        if (this._uiHandles) { return; }
-
-        var host = this.get(HOST),
-            bb = host.get(BOUNDING_BOX),
-            hide = Y.bind(host.hide, host),
-            uiHandles = [],
-            self = this,
-            hideOn = this.get('hideOn'),
-            i = 0,
-            o = {node: undefined, ev: undefined, keyCode: undefined};
-
-            //push all events on which the widget should be hidden
-            for (; i < hideOn.length; i++) {
-                
-                o.node = hideOn[i].node;
-                o.ev = hideOn[i].eventName;
-                o.keyCode = hideOn[i].keyCode;
-
-                //no keycode or node defined
-                if (!o.node && !o.keyCode && o.ev) {
-                    uiHandles.push(bb.on(o.ev, hide));
-                }
-
-                //node defined, no keycode (not a keypress)
-                else if (o.node && !o.keyCode && o.ev) {
-                    uiHandles.push(o.node.on(o.ev, hide));
-                }
-
-                //node defined, keycode defined, event defined (its a key press)
-                else if (o.node && o.keyCode && o.ev) {
-                    uiHandles.push(o.node.on(o.ev, hide, o.keyCode));
-                }
-                
-                else {
-                    Y.Log('The event with name "'+o.ev+'" could not be attached.');
-                }
-                
+            if (this.get(RENDERED)) {
+                this._bindUIAutohide();
+                this._syncUIAutohide();
             }
+        },
 
-        this._uiHandles = uiHandles;
-    },
+        destructor : function () {
 
-    _detachUIHandles : function () {
+            this._detachUIHandlesAutohide();
+        },
 
-        Y.each(this._uiHandles, function(h){
-            h.detach();
-        });
-        this._uiHandles = null;
-    },
+        _bindUIAutohide : function () {
 
-    _afterHostVisibleChange : function (e) {
+            this.after(VISIBLE+CHANGE, this._afterHostVisibleChangeAutohide);
+        },
 
-        this._uiSetHostVisible(e.newVal);
-    }
+        _syncUIAutohide : function () {
 
-}, {
+            this._uiSetHostVisibleAutohide(this.get(VISIBLE));
+        },
 
-    // *** Static *** //
+        // *** Private Methods *** //
 
-    NS : AUTOHIDE,
+        _uiSetHostVisibleAutohide : function (visible) {
 
-    ATTRS : {
+            if (visible) {
+                this._attachUIHandlesAutohide();
+                //Y.later(1, this, '_attachUIHandlesAutohide');
+            } else {
+                this._detachUIHandlesAutohide();
+            }
+        },
 
-       /*
-        * @description An array of events that will cause the widget to hide.
-        * Each index in the array should be an object literal with the following properties:
-        * 
-        * eventName (required, string): Refers to the event to listen to. If no other properties are 
-        * provided, it listens to this event on the widget's boundingBox. (ex: "clickoutside", "focusoutside")
-        *
-        * node (optional, Y.Node): Refers to the node on which the "eventName" event should be listening. 
-        * For example, to close the widget when a resize on "#someDiv" occurs, pass in "Y.one("#someDiv")" to node, and "resize:end" to eventName
-        *
-        * keyCode (optional, string): When listening to "key" events, this property can be filled with a keyCode such as "esc" or "down:27"
-        * 
-        * By default, this attribute has 3 objects within its array, clicking outside the widget, focussing outside the widget, and pressing the escape key
-        * @attribute hideOn
-        * @public
-        * @type array
-        */
-        hideOn: {
-            value: [
-                {
-                    eventName: CLICK_OUTSIDE
-                },
-                {
-                    eventName: FOCUS_OUTSIDE
-                },
+        _attachUIHandlesAutohide : function () {
 
-                {
-                    node: Y.one(DOCUMENT),
-                    eventName: KEY,
-                    keyCode: PRESS_ESCAPE
+            if (this._uiHandlesAutohide) { return; }
+
+            var bb = this.get(BOUNDING_BOX),
+                hide = Y.bind(this.hide,this),
+                uiHandles = [],
+                self = this,
+                hideOn = this.get('hideOn'),
+                i = 0,
+                o = {node: undefined, ev: undefined, keyCode: undefined};
+
+                //push all events on which the widget should be hidden
+                for (; i < hideOn.length; i++) {
+                    
+                    o.node = hideOn[i].node;
+                    o.ev = hideOn[i].eventName;
+                    o.keyCode = hideOn[i].keyCode;
+
+                    //no keycode or node defined
+                    if (!o.node && !o.keyCode && o.ev) {
+                        uiHandles.push(bb.on(o.ev, hide));
+                    }
+
+                    //node defined, no keycode (not a keypress)
+                    else if (o.node && !o.keyCode && o.ev) {
+                        uiHandles.push(o.node.on(o.ev, hide));
+                    }
+
+                    //node defined, keycode defined, event defined (its a key press)
+                    else if (o.node && o.keyCode && o.ev) {
+                        uiHandles.push(o.node.on(o.ev, hide, o.keyCode));
+                    }
+                    
+                    else {
+                        Y.Log('The event with name "'+o.ev+'" could not be attached.');
+                    }
+                    
                 }
-            ],
-            validator: Y.Lang.isArray
+
+            this._uiHandlesAutohide = uiHandles;
+        },
+
+        _detachUIHandlesAutohide : function () {
+
+            Y.each(this._uiHandlesAutohide, function(h){
+                h.detach();
+            });
+            this._uiHandlesAutohide = null;
+        },
+
+        _afterHostVisibleChangeAutohide : function (e) {
+
+            this._uiSetHostVisibleAutohide(e.newVal);
         }
-    }
+}
 
-});
 
-Y.namespace("Plugin").Autohide = WidgetAutohide;
+Y.WidgetAutohide = WidgetAutohide;